Change StudentRepository ID type from Integer to Long

+++ MiniProject/src/main/java/com/springboot/MiniProject/StudentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface StudentRepository extends JpaRepository<StudentDetails,Integer>{
+public interface StudentRepository extends JpaRepository<StudentDetails,Long>{
